backend/requirements.txt
Flask==3.1.1
openai==1.98.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.13
opencv-python==4.12.0.88
flask_sqlalchemy==3.1.1
Flask-Migrate==4.1.0

psycopg2-binary==2.9.10